# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_findMe",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/search": {
      "get": {
        "operationId": "UsersController_searchAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/skills": {
      "get": {
        "operationId": "SkillsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "skills"
        ]
      },
      "post": {
        "operationId": "SkillsController_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDto"
                }
              }
            }
          }
        },
        "tags": [
          "skills"
        ]
      }
    },
    "/skills/{id}": {
      "get": {
        "operationId": "SkillsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDto"
                }
              }
            }
          }
        },
        "tags": [
          "skills"
        ]
      },
      "patch": {
        "operationId": "SkillsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSkillDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDto"
                }
              }
            }
          }
        },
        "tags": [
          "skills"
        ]
      },
      "delete": {
        "operationId": "SkillsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDto"
                }
              }
            }
          }
        },
        "tags": [
          "skills"
        ]
      }
    },
    "/skills/search": {
      "get": {
        "operationId": "SkillsController_searchAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchSkillDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "skills"
        ]
      }
    },
...
  },
  "info": {
    "title": "Profile Info",
    "description": "Api de informações pessoais feita em NestJS com Prisma",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birth": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "example": [
              "USER",
              "ADMIN"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "birth",
          "description",
          "email",
          "password",
          "nickname",
          "phone",
          "picture",
          "address",
          "postal",
          "title",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginationOutputDto": {
        "type": "object",
        "properties": {
          "totalRecords": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          },
          "perPage": {
            "type": "number"
          },
          "nextPage": {
            "type": "number"
          },
          "prevPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "totalRecords",
          "currentPage",
          "perPage",
          "totalPages"
        ]
      },
      "PaginatedResponseDtoClass": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExperienceDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationOutputDto"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "SearchUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "birth": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "birth": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "birth",
          "description",
          "email",
          "password",
          "nickname",
          "phone",
          "picture",
          "address",
          "postal",
          "title"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "birth": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SkillDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ],
            "example": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "createdAt",
          "updatedAt",
          "userId"
        ]
      },
      "SearchSkillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ],
            "example": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ]
          }
        }
      },
      "CreateSkillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ],
            "example": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ]
          }
        },
        "required": [
          "name",
          "category"
        ]
      },
      "UpdateSkillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ],
            "example": [
              "LANGUAGE",
              "LIBRARY",
              "FRAMEWORK",
              "DATABASE",
              "TOOL",
              "SYSTEM",
              "TESTING",
              "PATTERN",
              "DEVOPS",
              "SOFT",
              "OTHER"
            ]
          }
        }
      },
...
      "ExperienceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "organization",
          "description",
          "start",
          "createdAt",
          "updatedAt",
          "userId"
        ]
      },
...
    }
  }
}